feat(monaco-editor): refactor Monaco editor structure and enhance functionality
- Introduced separate components for MonacoEditor, FallbackEditor, and SimpleTextEditor to improve code maintainability and performance.
- Added constants for CDN URLs and load attempts to streamline editor loading logic.
- Implemented optional chaining and nullish coalescing for safer property access and default values.
- Styled components using Tailwind CSS for a consistent user interface.
- Enhanced the editor's error handling and loading mechanisms for better user experience.
