+++ packages/ai-workspace-common/src/modules/artifacts/code-runner/monaco-editor/types.ts
@@ -1,0 +1,11 @@
+import { CodeArtifactType } from '@refly/openapi-schema';
+
+export interface MonacoEditorProps {
+  content: string;
+  language: string;
+  type: CodeArtifactType;
+  readOnly?: boolean;
+  isGenerating?: boolean;
+  canvasReadOnly?: boolean;
+  onChange?: (value: string) => void;
+}
